bank/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.22
Jinja2==2.11.2
pyotp==2.3.0
Werkzeug==1.0.1
MarkupSafe==1.1.1
itsdangerous==1.1.0
pyotp
uWSGI
requests
mysqlclient
PyJWT
xmltodict
lxml
setuptools
setuptools-rust
wheel